requirements.txt
flask
nats-py
asyncio